созданы новые компоненты для карточек, добавлена валидация формы
