+++ src/components/UIComponents/Input/Input.module.scss
@@ -1,14 +1,12 @@
 @import '../../../styles/variables.scss';
 
 .inputWrapper {
-  padding-left: 5px;
   display: flex;
   flex-direction: column;
   color: $text-color-descr;
-  font-size: 16px;
   gap: 4px;
   label {
-    padding-left: 4px;
+    padding: 0 4px;
   }
   input {
     color: $text-color-primary;
@@ -22,4 +20,10 @@
       margin: 0;
     }
   }
+}
+
+.error {
+  color: $text-color-error;
+  font-size: 14px;
+  padding: 0 4px;
 }
