+++ src/components/pagesComponents/MainPage/MainPromo/MainPromo.module.scss
@@ -75,7 +75,6 @@
   &__sum {
     width: 130px;
   }
-
   &__tel {
     width: 342px;
   }
