+++ src/components/pagesComponents/MainPage/MainPromo/blocks/PromoInternet/index.js
@@ -1,0 +1,1 @@
+export { PromoInternet } from './PromoInternet';
