+++ src/components/pagesComponents/MainPage/MainPromo/blocks/PromoNumber/index.js
@@ -1,0 +1,1 @@
+export { PromoNumber } from './PromoNumber';
