+++ src/components/pagesComponents/MainPage/MainPromo/blocks/PromoPay/index.js
@@ -1,0 +1,1 @@
+export { PromoPay } from './PromoPay';
