+++ src/components/pagesComponents/MainPage/MainPromo/blocks/PromoShop/index.js
@@ -1,0 +1,1 @@
+export { PromoShop } from './PromoShop';
