+++ src/components/pagesComponents/MainPage/MainPromo/blocks/PromoSlider/index.js
@@ -1,0 +1,1 @@
+export { PromoSlider } from './PromoSlider';
