+++ src/components/pagesComponents/MainPage/MainPromo/blocks/PromoTariff/index.js
@@ -1,0 +1,1 @@
+export { PromoTariff } from './PromoTariff';
