+++ src/styles/variables.scss
@@ -7,6 +7,7 @@ $text-color-secondary: $brand-color-primary;
 $text-color-tertiary: $brand-color-tertiary;
 $text-color-descr: #89909c;
 $text-color-copyright: #a1a5ae;
+$text-color-error: #ff3333;
 
 $bg-color-primary: $brand-color-tertiary;
 $bg-color-secondary: #f7f8fa;
